+++ mediation/AdColony/source/plugin/Assets/GoogleMobileAds/Mediation/AdColony/Api/AdColony/AdColonyMediationExtras.cs
@@ -14,7 +14,9 @@
 
 using System;
 
-namespace GoogleMobileAds.Api.Mediation.AdColony
+using GoogleMobileAds.Api.Mediation;
+
+namespace GoogleMobileAds.Mediation.AdColony.Api
 {
     public class AdColonyMediationExtras : MediationExtras
     {