+++ mediation/AdColony/source/plugin/Assets/GoogleMobileAds/Mediation/AdColony/Common/AdColony/DummyClient.cs
@@ -14,9 +14,9 @@
 
 using System.Reflection;
 using UnityEngine;
-using GoogleMobileAds.Api.Mediation.AdColony;
+using GoogleMobileAds.Mediation.AdColony.Api;
 
-namespace GoogleMobileAds.Common.Mediation.AdColony
+namespace GoogleMobileAds.Mediation.AdColony.Common
 {
     public class DummyClient : IAdColonyAppOptionsClient
     {